Tweaks for vision and autonomous

+++ robot/src/main/java/org/team2471/frc/steamworks/comm/VisionData.java
@@ -25,7 +25,7 @@ public class VisionData {
     return distance;
   }
 
-  public double getImageNumber() {
+  public int getImageNumber() {
     return imageNumber;
   }
